update:适配luatos camera api

+++ interface/src/luat_rtc_ec7xx.c
@@ -26,7 +26,7 @@
 #include "osasys.h"
 #include "luat_debug.h"
 #include "mw_aon_info.h"
-static int g_s_local_tz = 32;
+static int8_t g_s_local_tz = 32;
 int luat_rtc_set(struct tm *tblock){
     uint32_t Timer1 = (((tblock->tm_year+1900)<<16)&0xfff0000) | (((tblock->tm_mon+1)<<8)&0xff00) | ((tblock->tm_mday)&0xff);
     uint32_t Timer2 = ((tblock->tm_hour<<24)&0xff000000) | ((tblock->tm_min<<16)&0xff0000) | ((tblock->tm_sec<<8)&0xff00) | g_s_local_tz;
